llm/requirements.txt
torch
transformers
fastapiuvicorn